Added the "Embed()" function and fixed some bugs
- Add the optional "$base_dir" parameter to the class constructor. This
way we can use only the "$base_dir" without defining a virtual
"$template_file" for sections.
- Change the "NewSection()" function definition to a more logical way by
creating the new "Megan" object using only the "$base_dir" instead of
the incorrect "$template_file".
- Add the "SetTemplate()" function to define a section's template
instead of passing the template to its "Generate()" function. It's also
usable for the user.
- Add the "$is_file" parameter to the "Generate()" function to define if
the "$template" parameter is a file or not and automatically get the
file content if needed.
- Change the dynamic includes procedure to use the new "Generate()"
function feature and give the file's path instead of the file contents.
- Add the "Embed()" function to be able to put a file content or a
string to a global or local label and parse it as it's part of the
template.
- Fix the bug when a global label is not defined and Megan gets into an
infinite loop.